mise à jour Gladiator + Ajout class staic alert

+++ Gladiator/Equipments/Defend/Helmet.cs
@@ -8,7 +8,7 @@ namespace Equipments.Defend
 		private double _luckyParry;
 
 		#region IDefend implementation
-		public float LuckyParry {
+		public double LuckyParry {
 			get {
 				return this._luckyParry;
 			}
